requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.12
joblib==1.3.2
scikit-learn==1.3.2
pandas==2.2.2
numpy
openai==0.27.8
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1
azure-ai-formrecognizer
